Deployer: Fix integer overflow in memory computation.

+++ java/deployer/src/main/java/com/vmware/photon/controller/deployer/dcp/entity/ContainerTemplateService.java
@@ -64,7 +64,7 @@ public class ContainerTemplateService extends StatefulService {
     @NotNull
     @Immutable
     @Positive
-    public Integer memoryMb;
+    public Long memoryMb;
 
     /**
      * This value lists the amount of disk which should be allocated for the container, in GB.
